+++ .travis.yml
@@ -35,7 +35,6 @@ install:
 
 script:
   - ./morula all bin/spec_ci
-  - bin/spec
 
 
 notifications:
