--- bin/spec
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-set -e
-
-bin/lint
-bin/build
-go test
