--- exo-clean/.gitignore
@@ -1,2 +1,0 @@
-tmp/
-bin/exo-clean
