--- exo-clean/bin/setup
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-
-go get -u gopkg.in/alecthomas/gometalinter.v1
-gometalinter.v1 --install
-bin/build
