--- exo-clean/bin/spec
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-set -e
-
-bin/lint
-go test
