--- exo-create/README.md
@@ -1,4 +1,0 @@
-# "exo create" command
-
-This repo contains the code for the "exo create" command.
-This is part of the [Exosphere SDK](https://github.com/Originate/exosphere-sdk)
