--- exo-create/bin/build
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-set -e
-
-go build -o bin/exo-create
