--- exo-create/bin/spec.cmd
@@ -1,3 +1,0 @@
-call node_modules\o-tools-livescript\bin\build
-call node_modules\o-tools\bin\lint
-go test
