--- exo-create/bin/spec_ci
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-set -e
-
-bin/spec
