+++ exo-go/bin/setup
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+go get -u gopkg.in/alecthomas/gometalinter.v1
+gometalinter.v1 --install
