+++ exo-go/features/help.feature
@@ -1,0 +1,43 @@
+Feature: help command
+
+  As a developer not familiar with how to call "exo clean" correctly
+  I want to be given helpful information how to use this application
+  So that I can do the things I intended without having to look this up somewhere else.
+
+
+  Scenario: the user enters 'exo help'
+    When running "exo help" in the terminal
+    Then I see:
+      """
+      Usage:
+        exo [command]
+
+      Available Commands:
+        clean       Removes dangling Docker images and volumes.
+        create      Create a new Exosphere application
+        version     Exosphere go version number
+
+      Use "exo [command] --help" for more information about a command.
+      """
+
+
+  Scenario: the user enters 'exo help clean'
+    When running "exo help clean" in the terminal
+    Then I see:
+      """
+      Removes dangling Docker images and volumes.
+
+      Usage:
+        exo clean
+      """
+
+
+  Scenario: the user enters 'exo help create'
+    When running "exo help create" in the terminal
+    Then I see:
+      """
+      Create a new Exosphere application
+
+      Usage:
+        exo create
+      """
