+++ exo-go/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/Originate/exosphere/exo-go/src/cmd"
+
+func main() {
+	cmd.Execute()
+}
